Run interaction callbacks after cold starts finish
When a notification is tapped and the app is not backgrounded; it's not
even launched, the callback code is run before the router is ready to
navigate anywhere, which is a super common thing to want to do in an
interaction callback. This commit ensures the cold start is complete by
checking the segments in expo-router before running the interaction
callback (so that routing updates will run properly).

We also removed the tests... Not sure what great value they offer here.
They're super unit test-y and not very readable.